pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "deep-translator>=1.11.4",
    "gtts>=2.5.4",
    "nltk>=3.9.1",
    "pillow>=11.1.0",
    "pytesseract>=0.3.13",
    "streamlit>=1.43.2",
    "sumy>=0.11.0",
]
